alternator: deployment of alternator enabled
Deployment of Alternator is no supported by way of the alternator
object that is part of the cluster spec. The object has a port
as it's only property and this will be used to bind on.

If it is absent the operator will fall back to deploy a regular
Scylla cluster.